requirements.txt
# Python version requirement
# Requires Python 3.9 or higher
python_version >= "3.9"

# Core web framework
fastapi==0.115.2
uvicorn[standard]==0.30.6

# Template engine and web utilities
jinja2==3.1.4
python-multipart==0.0.12

# Content processing
markdown==3.7
pyyaml==6.0.2

# Jupyter notebook support
nbconvert==7.16.4
jupyter==1.1.1
jupyterlab==4.3.0
notebook==7.2.2

# Internationalization
babel==2.16.0

# Development utilities
watchfiles==0.24.0

# Additional dependencies for robust operation
aiofiles==24.1.0
python-dotenv==1.0.1

# Security and validation
email-validator==2.2.0

# Performance and caching
httpx==0.27.2
requests==2.31.0